auth

+++ client/src/components/Workout.jsx
@@ -5,10 +5,23 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 
 const Workout = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useWorkoutsContext;
+
   const handleDelete = async () => {
-    const response = await fetch(`https://gym-serv.onrender.com/api/workouts/${workout._id}`, {
-      method: "DELETE",
-    });
+    if (!user) {
+      return;
+    }
+
+    const response = await fetch(
+      `https://gym-serv.onrender.com/api/workouts/${workout._id}`,
+      {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${user.token}`,
+        },
+      }
+    );
     const data = await response.json();
     if (response.ok) {
       dispatch({ type: "DELETE_WORKOUT", payload: data });
